requirements.txt
fastapi==0.103.1
orjson==3.9.7
Pillow==10.0.1
python-dotenv==1.0.0
qrcode==7.4.2
uvicorn==0.23.2
